Apply PR suggestions: move DialogExtensions to extensions package, merge DialogExtensions functions into one (setStatusBarAsSurfaceColor)
