+++ WordPress/src/main/java/org/wordpress/android/util/WPActivityUtils.java
@@ -25,6 +25,7 @@ import com.google.android.material.appbar.MaterialToolbar;
 
 import org.wordpress.android.R;
 import org.wordpress.android.util.AppLog.T;
+import org.wordpress.android.util.extensions.DialogExtensionsKt;
 import org.wordpress.android.util.extensions.WindowExtensionsKt;
 
 import java.util.ArrayList;
