+++ WordPress/src/main/java/org/wordpress/android/util/extensions/DialogExtensions.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.util
+package org.wordpress.android.util.extensions
 
 import android.annotation.SuppressLint
 import android.app.Dialog
@@ -6,8 +6,6 @@ import android.os.Build
 import android.view.View
 import org.wordpress.android.R
 import org.wordpress.android.R.attr
-import org.wordpress.android.util.extensions.getColorFromAttribute
-import org.wordpress.android.util.extensions.isDarkTheme
 
 @SuppressLint("InlinedApi")
 fun Dialog.getPreferenceDialogContainerView(): View? {
@@ -28,13 +26,10 @@ fun Dialog.getPreferenceDialogContainerView(): View? {
 }
 
 fun Dialog.setStatusBarAsSurfaceColor() {
-    window?.statusBarColor = context.getColorFromAttribute(attr.colorSurface)
-}
-
-fun Dialog.updateSystemUiVisibility() {
-    window?.let {
+    window?.apply {
+        statusBarColor = context.getColorFromAttribute(attr.colorSurface)
         if (!context.resources.configuration.isDarkTheme()) {
-            it.decorView.systemUiVisibility = it.decorView
+            decorView.systemUiVisibility = decorView
                     .systemUiVisibility or View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
         }
     }
